% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contractions}
\alias{contractions}
\title{Contractions.}
\format{A data frame with two variables:
\describe{
\item{\code{key}}{Character string containing the contractions.}
\item{\code{value}}{Character string containing the long form of the contraction.}
}}
\usage{
contractions
}
\description{
American English language contractions and their long forms.
}
\keyword{datasets}
