% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoticons}
\alias{emoticons}
\title{Emoticons.}
\format{A data frame with 1 variable:
\describe{
\item{\code{key}}{Character string containing the symbols that make up the emoticon.}
}}
\usage{
emoticons
}
\description{
Common emoticons used on the web and in messaging applications.
}
\keyword{datasets}
