% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internetAbbreviations}
\alias{internetAbbreviations}
\title{Internet Abbreviations.}
\format{A data frame with two variables:
\describe{
\item{\code{key}}{Character string containing the abbreviated text.}
\item{\code{value}}{Character string containing the long form of the abbreviation.}
}}
\usage{
internetAbbreviations
}
\description{
Common abbreviations used on the internet and in messaging platforms.
}
\keyword{datasets}
