% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profanity}
\alias{profanity}
\title{Profanity}
\format{A data frame:
\describe{
\item{\code{key}}{Character string containing the offending text.}
}}
\source{
\url{https://www.freewebheaders.com/full-list-of-bad-words-banned-by-google/}
}
\usage{
profanity
}
\description{
A list of words considered offensive for internet use.
}
\keyword{datasets}
